choozi - Extract drawing of arcs into lib for others to use
